pyproject.toml
# poetry config
[tool.poetry]
name = "server"
version = "0.1.0"
description = ""
authors = ["discoboy <jyjyartem@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
fastapi = {extras = ["all"], version = "^0.103.1"}
sqlalchemy = "^2.0.21"
asyncpg = "^0.28.0"
bcrypt = "^4.0.1"
redis = "^5.0.0"
pyjwt = "^2.8.0"
fastapi-mail = "^1.4.1"
pytz = "^2023.3.post1"
pillow = "^10.0.1"
boto3 = "^1.28.53"
flake8-bugbear = "^23.9.16"
flake8-builtins = "^2.1.0"
alembic = "^1.12.0"
celery = "^5.3.4"
flower = "^2.0.1"

[tool.poetry.group.dev.dependencies]
pytest = "^7.4.2"
pytest-asyncio = "^0.21.1"
pytest-dotenv = "^0.5.2"
pyparsing = "^3.1.1"
flake8 = "^6.1.0"
flake8-pyproject = "^1.2.3"
flake8-commas = "^2.1.0"
flake8-quotes = "^3.3.2"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"


# pytest config
[tool.pytest.ini_options]
pythonpath = [".", "src"]
env_files = ["test.env"]
asyncio_mode = "auto"
